Refatoração parcial

+++ src/main/java/br/com/myhomefinances/domain/TipoConta.java
@@ -13,21 +13,21 @@ public class TipoConta implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 
 	public TipoConta() {}
 
-	public TipoConta(Integer id, String nome) {
+	public TipoConta(Long id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
